Include more metadata for non-posts
